+++ resolvers/index.js
@@ -1,6 +1,6 @@
 const { Constructor, Constructors } = require("../data/Constructors/resolvers");
 const { Driver, Drivers } = require("../data/Drivers/resolvers");
-const { SeasonsResolver } = require("./Seasons");
+const { Seasons } = require("../data/Seasons/resolvers");
 const { StatusResolver } = require("../data/Status/resolvers");
 const { Results } = require("../data/Results/resolvers");
 const { QualifyingResults } = require("../data/Qualifying/resolvers");
@@ -14,7 +14,7 @@ const resolvers = {
     Driver,
     Drivers,
     Results,
-    Seasons: SeasonsResolver,
+    Seasons,
     Status: StatusResolver,
     QualifyingResults,
     DriverStandings,
